split profiling test into separate xctestcase
