+++ Samples/iOS-Swift/iOS-SwiftUITests/UITestHelpers.swift
@@ -1,0 +1,12 @@
+import XCTest
+
+extension XCUIElement {
+    func waitForExistence(_ message: String) {
+        XCTAssertTrue(self.waitForExistence(timeout: TimeInterval(10)), message)
+    }
+
+    func afterWaitingForExistence(_ failureMessage: String) -> XCUIElement {
+        waitForExistence(failureMessage)
+        return self
+    }
+}
